+++ api/src/main/java/net/jqwik/api/Functions.java
@@ -50,8 +50,15 @@ public class Functions {
 		}
 
 		/**
+		 * @deprecated Use {@linkplain #returning(Arbitrary)} instead. Will be removed in 1.7.0.
+		 */
+		@API(status = DEPRECATED, since = "1.6.0")
+		public <F, R> FunctionArbitrary<F, R> returns(Arbitrary<R> resultArbitrary) {
+			return returning(resultArbitrary);
+		}
+
+		/**
 		 * Create an arbitrary to create instances of functions represented by this wrapper.
-		 *
 		 * The generated functions are guaranteed to return the same result
 		 * given the same input values.
 		 *
@@ -62,7 +69,8 @@ public class Functions {
 		 * @param <R> The return type of the functional interface
 		 * @return a new arbitrary instance
 		 */
-		public <F, R> FunctionArbitrary<F, R> returns(Arbitrary<R> resultArbitrary) {
+		@API(status = MAINTAINED, since = "1.6.0")
+		public <F, R> FunctionArbitrary<F, R> returning(Arbitrary<R> resultArbitrary) {
 			return FunctionsFacade.implementation.function(functionalType, resultArbitrary);
 		}
 	}